Utilize jumbo mbuf clusters in bsdmodule. Fix mbuf leak.
Enables processing and allocation of 2k-16k jumbo frames.
Also fixes allocation and deallocation problems that occurred
when modifying cloned packets of smaller sizes as well.

Note that mbufs are currently used quite suboptimally in Click.
One problem is the requirement of having contiguous data,
and another issue is that data needs to be copied into a new
cluster e.g. when a cloned packet is modified.