Bsdmodule: Rename the "new" makefile so that it's not used by accident.
Using a makefile that includes the default options for compiling
kernel modules doesn't sound like a bad idea, but currently this
makefile doesn't work (e.g., due to C++ compilation issues). Hence we
can still keep it, but we rename it so that it's not used if somebody
happens to run BSD make in the bsdmodule directory.